+++ src/main.js
@@ -16,8 +16,10 @@ import store from '@/store';
 import VueI18n from 'vue-i18n';
 import Vuex from 'vuex';
 import VueQrcode from '@xkeshi/vue-qrcode';
-import PopOver from '@/components/PopOver';
+import Toasted from 'vue-toasted';
+import * as toastConfig from './toast.config';
 
+import PopOver from '@/components/PopOver';
 import StandardButton from '@/components/Buttons/StandardButton';
 import StandardInput from '@/components/StandardInput';
 
@@ -64,7 +66,17 @@ Vue.component('standard-input', StandardInput);
 Vue.config.productionTip = false;
 
 Vue.use(Vuex);
+Vue.use(Toasted);
 Vue.use(VeeValidate);
+
+// Register global toasts
+Object.keys(toastConfig).forEach(item => {
+  Vue.toasted.register(
+    toastConfig[item].name,
+    toastConfig[item].payloadFunc,
+    toastConfig[item].options
+  );
+});
 
 /* Init Bootstrap */
 Vue.use(BootstrapVue);